refactor: move subComponents below IdeationWrapper
