+++ src/app/(main)/my-voyage/[teamId]/ideation/components/dependencies/FinalizedIdeationCard.tsx
