+++ src/app/(main)/my-voyage/[teamId]/ideation/components/dependencies/VoteCard.tsx
