+++ src/app/(main)/my-voyage/[teamId]/ideation/components/dependencies/VoteDescriptionCard.tsx
