+++ src/app/(main)/my-voyage/[teamId]/ideation/page.tsx
@@ -1,6 +1,6 @@
+import Spinner from "@/components/Spinner";
 import { Suspense } from "react";
 import IdeationComponentWrapper from "./components/IdeationComponentWrapper";
-import Spinner from "@/components/Spinner";
 
 interface IdeationPageProps {
   params: {
